refactor: ChallengeEvent 핸들링 로직 분리
